requirements.txt
######-Requirements-######

#-Pwntools-#
pwntools
#-Colorama-#
colorama
#-Requests-#
requests
